requirements.txt
rhasspy-hermes==0.1.5
paho-mqtt==1.5.0
PyAudio==0.2.11
